+++ R/landisbc-package.R
@@ -6,11 +6,11 @@ utils::globalVariables(c(
   ".data", "Age", "age", "Area", "bark", "bcwsft", "bec_group", "bec_zone", "bole", "branches",
   "BURN_SEVERITY_RATING", "CleanMapCode", "CLSTR_ID", "DATE", "DAY", "EcoCode",
   "exclude_bec_label", "exclude_tsa", "FIRE_NUMBER", "FIRE_YEAR", "foliage", "Freq",
-  "FUEL_TYPE_CD", "hash", "INTERSECTS", "JULIAN_DAY", "lev2", "M1_2_PERCENT_CONIFER", "MapCode",
-  "min_leading_pct", "MONTH", "MRSRD_A", "MRSRD_Y", "PERCENT_CONIFER", "PERCENT_DEAD_FIR",
-  "planting_sph", "provincial", "series_id", "SITE_IDENTIFIER", "site_index", "SIZE_HA",
-  "SPECIES", "species", "species_group", "SpeciesCode", "tipsy_species", "total", "TotalArea",
-  "value", "YEAR"
+  "FUEL_TYPE_CD", "hash", "include_bec_labels", "INTERSECTS", "JULIAN_DAY", "lev2",
+  "M1_2_PERCENT_CONIFER", "MapCode", "min_leading_pct", "MONTH", "MRSRD_A", "MRSRD_Y",
+  "PERCENT_CONIFER", "PERCENT_DEAD_FIR", "planting_sph", "provincial", "series_id",
+  "SITE_IDENTIFIER", "site_index", "SIZE_HA", "SPECIES", "species", "species_group",
+  "SpeciesCode", "tipsy_species", "total", "TotalArea", "value", "YEAR"
 ))
 
 #' @keywords internal